Error handling

+++ src/main/java/com/holis/san01/model/Entidade.java
@@ -2,7 +2,6 @@ package com.holis.san01.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import org.springframework.lang.Nullable;
 
 import javax.validation.constraints.Pattern;
 import java.util.Date;
@@ -14,62 +13,114 @@ public class Entidade {
     @Id
     @Column(nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long codEntd;
+    private int codEntd;
 
     @Column(nullable = false)
     private String nome;
+
     @Column(nullable = false)
     private String razaoSocial;
+
     private String indCliente;
+
     private String indFornec;
+
     private String indTransp;
+
     private String tpPessoa;
+
     private String cgc;
+
     private String rg;
+
     private String logradouro;
+
     private String numero;
+
     private String complemento;
+
     private String bairro;
+
     private String localidade;
+
     private String estado;
+
     private String pais;
+
     private String cep;
+
     private String contato;
+
     private String email;
+
     private String telefone;
+
     private String caixaPostal;
+
     private String inscrEstadual;
+
     private String inscrMunicipal;
+
     private Integer indPagto;
+
     private Integer vencFeriado;
+
     private Integer vencDomingo;
+
     private Integer venSabado;
+
     private String homePage;
+
     private Boolean logPossuiNfEletro;
+
     private Boolean logNfEletro;
+
     private String codCondPag;
+
     private String natOperSai;
+
     private String natOperEnt;
+
     private String cgcCob;
+
     private String cepCob;
+
     private String estadoCob;
+
     private String cidadeCob;
+
     private String bairroCob;
+
     private String enderecoCob;
+
     private String paisCob;
+
     private String cxPostCob;
+
     private String insEstCob;
+
     private Integer codMensagem;
+
     private String emailNfe;
+
     private Integer tipoFinIdSai;
+
     private Integer tipoFinIdEnt;
+
     private String setor;
+
     private String segmento;
+
     private String ramo;
+
     private String obsEntrega;
+
     private String observacoes;
+
     private String situacao;
+
     private Date dtCriacao;
+
     @Column(nullable = false)
     @Pattern(regexp = "[SN]", message = "O campo 'archive' deve ser 'S' ou 'N'.")
     private String archive;
